task_5 release, task_7 start
Fixed all memory issues in task_5.

Added a new "memory_manager" prototype for task_7,
assuming the presence of a combination of MT + Intrinsics + MM in the release.

I've been assigned to task_8 and plan to use
Python with CUDA or similar technologies to accomplish it.
It's likely that the solution to this problem will be available
here before task_7.